Move DbUtils implemetation to separate package.
