+++ testing/db-testkit/src/main/kotlin/net/corda/db/testkit/dbutilsimpl/DbUtilsHelper.kt
@@ -1,4 +1,4 @@
-package net.corda.db.testkit
+package net.corda.db.testkit.dbutilsimpl
 
 import com.typesafe.config.Config
 import net.corda.db.core.CloseableDataSource
