+++ testing/db-testkit/src/main/kotlin/net/corda/db/testkit/dbutilsimpl/HSQLHelper.kt
@@ -1,10 +1,11 @@
-package net.corda.db.testkit
+package net.corda.db.testkit.dbutilsimpl
 
 import com.typesafe.config.Config
 import com.typesafe.config.ConfigFactory
 import com.typesafe.config.ConfigValueFactory
 import net.corda.db.core.CloseableDataSource
 import net.corda.db.core.createDataSource
+import net.corda.db.testkit.TestInMemoryEntityManagerConfiguration
 import net.corda.orm.EntityManagerConfiguration
 import net.corda.schema.configuration.DatabaseConfig
 import net.corda.test.util.LoggingUtils.emphasise
